reorganize types for nodepool
